Make TimeTravel "Sticky"
This commit makes the Time Travel feature "stick" until it is cleared with the "reset" link in the Time Travel status bubble.

This is useful for going back in time and performing multiple tasks, such as searching for a station, filtering on one station, then filtering on a different station.  Previously each of these operations required going back into the Time Travel dialog to set a new Time Travel date.