Minor Optimizations
Added a little javascript to hide the "flash-frame" about page that comes up for browsers with javascript disabled.  This will have no effect if JS is disabled and the about page will still display.

A couple tweaks to external javascript loading.

Moved some js to main.js to speed loading.