Performance & UI Enhancements
This is a first pass at improving performance of the station details (overview.php) window.  In some cases, the Packet Frequency calculations can take longer than a second alone causing the spinner to appear for 2-3 seconds.  This has been moved to an ajax call, so the details window loads in less than a second in most instances with the packet frequency data loading immediately after.

Wrapped weather data in a fieldset for the mini-popup window to isolate it from the other station details.