Raw Packet History & Leaflet Optimization
With the previous addition of days_to_save_packet_data in the database config, this commit allows raw packet history to be filtered by date range, up to the number of days defined in days_to_save_packet_data in the config.

days_to_save_packet_data defines the period in which raw packets are scrubbed from the database to save space.

Additionally, leaflet now has preferCanvas set to true, which is recomended for performance on maps with many markers.  (Disclaimer: I have not noticed any significant difference in performance)