"Live" is the way to live!
This is a significant update...

APRS by definition is a live, realtime messaging system for transporting information.  This currently exists in the map view, not to the static displays of station information.  This release fixes that, and then some...

The station details window now has several live data elements..  If a packet comes in while one of these windows is open, some or all of the elements in the view will update automatically.

This includes;

1. Station Overview: (last packet, last position, comment, date/time, path, etc.).  Data in fields will gently flash anytime an update occurs.

2. Weather: Current Conditions gauges update with animation anytime a new weather packet is received.

3. Telemetry: Current Readings tables update whenever a new telemetry packet is received.

A gorgeous new Live Packet Feed console is available to view the raw packets for a station as the arrive in real-time.

The Raw Packet view has also been updated with the same "console" appearance as the Live Packet Feed for a consistent user experience.  A streamlined search form and pagination function make better use of available screen realestate.

Fixes include; instances where links in the station search results force a resource-intensive full page load, instead of loading via ajax.

This release also includes some minor css styling fixes